flink-job/requirements.txt
apache-flink>=1.20.0  # or the latest version
cassandra-driver
